Move helpers into LeoUnitTest class where they belong.
