+++ leo/unittests/commands/test_gotoCommands.py
@@ -2,14 +2,13 @@
 #@+node:ekr.20230802060212.1: * @file ../unittests/commands/test_gotoCommands.py
 """Tests of leo.commands.gotoCommands."""
 from leo.core import leoGlobals as g
-# from leo.core.leoTest2 import LeoUnitTest
+from leo.core.leoTest2 import LeoUnitTest
 from leo.commands.gotoCommands import GoToCommands
-from leo.unittests.commands.test_outlineCommands import TestOutlineCommands
 assert g
 
 #@+others
-#@+node:ekr.20230802060212.2: ** class TestGotoCommands(TestOutlineCommands)
-class TestGotoCommands(TestOutlineCommands):
+#@+node:ekr.20230802060212.2: ** class TestGotoCommands(LeoUnitTest)
+class TestGotoCommands(LeoUnitTest):
     """Unit tests for leo/commands/gotoCommands.py."""
 
     #@+others
